# src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: Teensy
  description: Tiny Url Generator
  version: 1.0.0
  contact:
    name: Tejas Ozarkar
    email: tejozarkar@gmail.com
servers:
  - url: http://localhost:8080
    description: Local Environment
  - url: http://teensyapi-env.eba-k2xfeapv.ap-south-1.elasticbeanstalk.com
    description: Production Environment
paths:
...
  /group/admin/{groupId}/{userId}:
    post:
      tags:
        - group
      summary: Delete group admin
      operationId: deleteGroupAdmin
      responses:
        200:
          description: Get Group Admins
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      parameters:
        - name: groupId
          in: path
          description: Id of Group
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          description: Id of user
          required: true
          schema:
            type: integer
  /group/isadmin/{groupId}:
    get:
      tags:
        - group
      summary: Get group admins
      operationId: isGroupAdmin
      responses:
        200:
          description: Get Group Admins
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IsAdminDto'
      parameters:
        - name: groupId
          in: path
          description: Id of Group
          required: true
          schema:
            type: integer
  /group/admin/{groupId}:
    post:
      tags:
        - group
      summary: Get group admins
      operationId: addGroupAdmin
      requestBody:
        description: Create New Group
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminDto'
      responses:
        200:
          description: Get Group Admins
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      parameters:
        - name: groupId
          in: path
          description: Id of Group
          required: true
          schema:
            type: integer
    get:
      tags:
        - group
      summary: Get group admins
      operationId: getGroupAdmins
      responses:
        200:
          description: Get Group Admins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDto'
      parameters:
        - name: groupId
          in: path
          description: Id of Group
          required: true
          schema:
            type: integer
...
  /group/{groupId}:
    get:
      tags:
        - group
      summary: Get selected card
      operationId: getGroupInfo
      responses:
        200:
          description: Get Group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
      parameters:
        - name: groupId
          in: path
          description: Id of Group
          required: true
          schema:
            type: integer
    delete:
      summary: Delete selected card
      operationId: deleteGroup
      responses:
        200:
          description: Delete Group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      parameters:
        - name: groupId
          in: path
          description: Id of Group
          required: true
          schema:
            type: integer
  /group:
    post:
      summary: Crate New Group
      operationId: createGroup
      requestBody:
        description: Create New Group
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupDto'
        required: true
      responses:
        200:
          description: New Group Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
    get:
      summary: Get all groups of user
      operationId: getUserGroups
      responses:
        200:
          description: Get user groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupResponse'
...
components:
  schemas:
...
    AdminDto:
      type: object
      properties:
        email:
          type: string
...
    UserResponseDto:
      type: object
      properties:
        id:
          type: integer
        firstname:
          type: string
        lastname:
          type: string
        username:
          type: string
        email:
          type: string
...
    GroupDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
    GroupResponse:
      allOf:
        - $ref: '#/components/schemas/GroupDto'
      type: object
      properties:
        id:
          type: integer
    IsAdminDto:
      type: object
      properties:
        isadmin:
          type: boolean
    GenericResponse:
      type: object
      properties:
        message:
          type: string